requirements.txt
fastapi
uvicorn[standard]

onnxruntime-gpu==1.18.0

opencv-python-headless
numpy

mediapipe

joblib
scikit-learn
python-multipart

httptools
uvloop

starlette
